pyproject.toml
[project]
name = "data-canary"
version = "0.0.1"
description = "AI-powered data quality copilot – finds everything wrong with your data in plain English"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Jonathan Lam", email = "lamheihung@gmail.com"}
]

dependencies = [
    "streamlit>=1.38",        
    "pandas",                 
    "polars",          
    "python-dotenv>=1.0",     
    "pydantic>=2.7",
    "google-genai"        
]

[project.optional-dependencies]
dev = [
    "ruff",      
    "black",     
    "mypy",     
    "pytest"    
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"